handle the InAppCall Notification Dialog show or not as per settings, add two Notification type for redirect to Web browser and emergency alert or notice for Society , handle delete family member as per Role , Remove Vendor Name Display from expense list, handle if visitor check-in deny by user or Security

+++ lib/Activities/AlreadyPaid.dart
@@ -14,7 +14,7 @@ import 'base_stateful.dart';
 
 class BaseAlreadyPaid extends StatefulWidget {
   String invoiceNo;
-  int amount;
+  double amount;
   BaseAlreadyPaid(this.invoiceNo,this.amount);
 
   @override
@@ -60,7 +60,7 @@ class AlreadyPaidState extends BaseStatefulState<BaseAlreadyPaid> {
   bool isStoragePermission=false;
 
   String invoiceNo;
-  int amount;
+  double amount;
   AlreadyPaidState(this.invoiceNo,this.amount);
 
   @override
